+++ types/utils/database/sqlite.d.ts
@@ -1,0 +1,6 @@
+export default class _default extends DBClientClass {
+    constructor(prompts: any, optionsCDS: any);
+    listTables(): Promise<any>;
+    #private;
+}
+import DBClientClass from "./index.js";